+++ apps/marketplace/components/BuyerATM/BuyerATMAdditionalInformationStage.scss
